integration_tests: move write_nv_index() to common lib
